+++ core/src/main/kotlin/ca/allanwang/kau/logging/TimberLogger.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package ca.allanwang.kau.logging
 
 import timber.log.Timber
@@ -9,11 +11,11 @@ import timber.log.Timber
  * Timber extension that will embed the tag as part of the message for each log item
  */
 open class TimberLogger(tag: String) {
-    internal val TAG = "$tag: %s"
-    fun e(s: String) = Timber.e(TAG, s)
-    fun e(t: Throwable?, s: String = "error") = if (t == null) e(s) else Timber.e(t, TAG, s)
-    fun d(s: String) = Timber.d(TAG, s)
-    fun i(s: String) = Timber.i(TAG, s)
-    fun v(s: String) = Timber.v(TAG, s)
-    fun eThrow(s: String) = e(Throwable(s))
+    val TAG = "$tag: %s"
+    inline fun e(s: String) = Timber.e(TAG, s)
+    inline fun e(t: Throwable?, s: String = "error") = if (t == null) e(s) else Timber.e(t, TAG, s)
+    inline fun d(s: String) = Timber.d(TAG, s)
+    inline fun i(s: String) = Timber.i(TAG, s)
+    inline fun v(s: String) = Timber.v(TAG, s)
+    inline fun eThrow(s: String) = e(Throwable(s))
 }