+++ core/src/main/kotlin/ca/allanwang/kau/utils/Const.kt
@@ -1,14 +1,16 @@
 package ca.allanwang.kau.utils
 
+import android.support.v4.widget.ViewDragHelper
+
 /**
  * Created by Allan Wang on 2017-06-08.
  */
 const val ANDROID_NAMESPACE = "http://schemas.android.com/apk/res/android"
 
-const val KAU_LEFT = 1
-const val KAU_TOP = 2
-const val KAU_RIGHT = 4
-const val KAU_BOTTOM = 8
+const val KAU_LEFT = ViewDragHelper.EDGE_LEFT
+const val KAU_RIGHT = ViewDragHelper.EDGE_RIGHT
+const val KAU_TOP = ViewDragHelper.EDGE_TOP
+const val KAU_BOTTOM = ViewDragHelper.EDGE_BOTTOM
 const val KAU_HORIZONTAL = KAU_LEFT or KAU_RIGHT
 const val KAU_VERTICAL = KAU_TOP or KAU_BOTTOM
 const val KAU_ALL = KAU_HORIZONTAL or KAU_VERTICAL