+++ core/src/main/kotlin/ca/allanwang/kau/utils/PackageUtils.kt
@@ -1,6 +1,5 @@
 package ca.allanwang.kau.utils
 
-import android.annotation.SuppressLint
 import android.content.Context
 import android.content.pm.PackageManager
 import android.os.Build